fixed bigint[] && integer[] error

+++ test/unit/cache/triggers/commutative/sum/sum_where_includes_to_concat_two_arrays/public.orders.sql
@@ -2,12 +2,12 @@ create or replace function cache_totals_for_companies_on_orders()
 returns trigger as $body$
 declare matched_old boolean;
 declare matched_new boolean;
-declare inserted_clients_ids integer[];
-declare inserted_partners_ids integer[];
-declare not_changed_clients_ids integer[];
-declare not_changed_partners_ids integer[];
-declare deleted_clients_ids integer[];
-declare deleted_partners_ids integer[];
+declare inserted_clients_ids bigint[];
+declare inserted_partners_ids bigint[];
+declare not_changed_clients_ids bigint[];
+declare not_changed_partners_ids bigint[];
+declare deleted_clients_ids bigint[];
+declare deleted_partners_ids bigint[];
 begin
 
     if TG_OP = 'DELETE' then
